feat: update IBAN handling in MerchantUpdateIbanServiceImpl and add corresponding tests
