+++ src/test/java/it/gov/pagopa/merchant/controller/merchant_portal/MerchantPortalMerchantControllerImplTest.java
@@ -8,6 +8,7 @@ import it.gov.pagopa.merchant.constants.MerchantConstants.ExceptionCode;
 import it.gov.pagopa.merchant.constants.MerchantConstants.ExceptionMessage;
 import it.gov.pagopa.merchant.dto.InitiativeDTO;
 import it.gov.pagopa.merchant.dto.MerchantDetailDTO;
+import it.gov.pagopa.merchant.dto.MerchantIbanPatchDTO;
 import it.gov.pagopa.merchant.dto.ReportedUserCreateResponseDTO;
 import it.gov.pagopa.merchant.dto.ReportedUserDTO;
 import it.gov.pagopa.merchant.exception.custom.MerchantNotFoundException;
@@ -24,6 +25,7 @@ import org.springframework.boot.security.autoconfigure.UserDetailsServiceAutoCon
 import org.springframework.boot.webmvc.test.autoconfigure.AutoConfigureMockMvc;
 import org.springframework.boot.webmvc.test.autoconfigure.WebMvcTest;
 import org.springframework.context.annotation.Import;
+import org.springframework.http.MediaType;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
@@ -34,6 +36,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
@@ -108,7 +111,7 @@ class MerchantPortalMerchantControllerImplTest {
     }
 
     @Test
-    void getMerchantDetailByMerchantIdAndInitiativeId_notFound() throws Exception {
+  void getMerchantDetailByMerchantIdAndInitiativeId_notFound() throws Exception {
         Mockito.when(merchantServiceMock.getMerchantDetail(MERCHANT_ID, INITIATIVE_ID))
                 .thenReturn(null);
 
@@ -127,6 +130,28 @@ class MerchantPortalMerchantControllerImplTest {
         Assertions.assertEquals(ExceptionCode.MERCHANT_NOT_ONBOARDED, errorDTO.getCode());
         Assertions.assertEquals(String.format(ExceptionMessage.INITIATIVE_AND_MERCHANT_NOT_FOUND, INITIATIVE_ID),errorDTO.getMessage());
         Mockito.verify(merchantServiceMock).getMerchantDetail(anyString(), anyString());
+    }
+
+    @Test
+    void updateIban() throws Exception {
+        MerchantIbanPatchDTO requestDto = new MerchantIbanPatchDTO();
+        requestDto.setIban("IT60X0542811101000000123456");
+
+        MerchantDetailDTO responseDto = MerchantDetailDTOFaker.mockInstance(1);
+        Mockito.when(merchantServiceMock.patchMerchant(
+                anyString(), anyString(), any(MerchantIbanPatchDTO.class))
+        ).thenReturn(responseDto);
+
+        mockMvc.perform(
+                        patch("/idpay/merchant/portal/initiatives/{initiativeId}", INITIATIVE_ID)
+                                .header("x-merchant-id", MERCHANT_ID)
+                                .contentType(MediaType.APPLICATION_JSON)
+                                .content(objectMapper.writeValueAsString(requestDto))
+                )
+                .andExpect(status().isOk());
+
+        Mockito.verify(merchantServiceMock)
+                .patchMerchant(anyString(), anyString(), any(MerchantIbanPatchDTO.class));
     }
 
   @Test
